introduce new arrow utils for working with batches
